refactor(core): 优化扩展状态初始化和控制器设置逻辑
- 在 Category 和 Sentence 扩展中为 status 字段添加默认初始化，避免空指针
- 为 Status 相关字段赋予默认值，实现更合理的默认状态
- 移除 CategoryReconciler 和 SentenceReconciler 中重复的状态初始化方法和调用
- 简化 ControllerBuilder 配置，合并链式调用
- 新增 OverviewController 提供概览接口，返回示例数据
- 新增 SentenceController 支持批量创建句子，并返回统计结果
- 调整代码格式，优化注解多行写法以提升可读性

+++ src/main/java/top/puresky/hitokotohub/controller/SentenceController.java
@@ -49,8 +49,8 @@ public class SentenceController {
 
     @Data
     public static class BatchCreateSentenceResult {
-        private int total;
-        private int success;
-        private int failed;
+        private long total;
+        private long success;
+        private long failed;
     }
 }
